Add AlphaBanner component
Didn't make it configurable since we're not likely to be changing
it anytime soon.

Added a test that doesn't do much, and added a French translation,
but didn't configure it into the component since we don't really
have a good way to pass global things to `Layout` at this point.

Heavily cribbed from:
- https://design-system.service.gov.uk/components/phase-banner/