Set locale in express middleware
Instead of looking for query parameters in one specific route,
set the locale in the middleware instead, so that it runs no matter
which route you use.

Still setting a variable on the server (which is a bad idea), but
since we're not yet worried about en/fr text, we can solve this
problem later.

More on express middleware here:
- https://expressjs.com/en/guide/writing-middleware.html