Use smaller favicon file
The login page is ~6KB to download whereas the favicon file was ~15KB.

Can you believe?

This commit adds a 1KB 32x32 png favicon file and changes the
reference in the site's <head>.

It also deletes the old one, which is fine since this is more or
less a placeholder anyway.