Remove date of birth field from login form
We've previously talked about asking for a date of birth when
users sign in to see our service, and I put it in initially to
hue to that idea.

However, none of our stories actually include asking for a date of
birth, and doing a login form that asks for Date of Birth is harder
than doing one that doesn't ask for it.

Checked with Cayce and she says "drop it".

So dropping it.