Add hidden "Error: " text to validation errors
Also changed the error for the login page to make it a bit shorter.

Meant I had to update a bunch of unit tests.

This is best practice for screen readers, because someone using one
wouldn't be able to see the visual error cues (ie, that the text is
red or the red border).
- https://design-system.service.gov.uk/components/error-message/#how-it-works