First stab at adding a title to SummaryTable
After googling around somewhat, there seems to be no "official"
way to add a title to a description list.

There was some markup about adding a title attribute to the description
list so that's what I went with.
Otherwise, it looks like adding a header or a paragraph before your
list is the way to go.

Sources:
MDN for Description lists:
- https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dl

Example with a title:
- https://www.w3.org/WAI/WCAG21/Techniques/html/H40

Stackoverflow "adding a title to a list"
- https://stackoverflow.com/questions/1141639/how-to-semantically-provide-a-caption-title-or-label-for-a-list-in-html