Added 2 more tables to dashboard page
Busier, but, again, this seems to be the generic interface we're
going for.

Updated the test slightly to be more inclusive for where it looks
for the row values to be printed, and to make sure the h2s exist.

Also updated the prettier instructions to allow longer line lengths
otherwise, prettier and eslint clash overformatting.