Remove duplicate 'renderStylesToString'
When refactoring the renderPage function earlier, I accidentally
left an extra renderStylesToString call lying around.
(That is, it was getting called in 'renderPage' and '_renderDocument'.)

This didn't cause any practical issues, but it did duplicate all the
style tags in the HTML, unnecessarily sending more markup down the
wire.

Looks after removing it, response sizes are down ~25%.