Don't save name to session on login page
Only save a user to a session if a user has been found.
Saving a name allowed us to pass back in the session to the page
and see the name we just entered in the form field, but we could jsut
pass it in directly and bypass the session altogether.