Add Footer element with "Get help" link
The link doesn't go anywhere, but it was something that we wanted
to go in the Footer of our page.

Since we didn't have a footer, I created one.

Since footers are weird with variable content, I used this
cheap and easy CSS tactic to get it to stick to the bottom unless
the page content was larger than the height of the browser.
- https://medium.freecodecamp.org/how-to-keep-your-footer-where-it-belongs-59c6aa05c59c

Looks like we could also try flexbox if that was something we were
comfortable with.