Add 'cypress run' cli comment
Running 'cypress open' will pop open a browser and let you see the
actual tests as they move through the interface.

However, if we want to run these tests more quickly, and, particularly,
as part of our CD pipeline, we need a trimmer way to run our functional
tests.

So this should be able to run inside of GitHub actions.