Add pa11y testing to the first 3 URLs
This is a proof of concept PR that integrates Pa11y HTML a11y
scanning for the first three pages of the app:

- welcome page: "/"
- login page: "/login"
- introduction page: "/introduction"

It doesn't get everything, and it isn't really optimal for scanning
logged in pages, but it can scan them all with a bit of config.
For all logged-in pages, you would have to include the actions needed
to get to that page, which means a bit of overhead.

However, for the most part, this should be a set-it-and-forget-it thing.

Didn't use pa11y-ci because the documentation isn't as good, but
we could use that instead if we preferred.

Most of this code was lifted straight from what we did on the IRCC
rescheduler.

Sources:
- Pa11y: http://pa11y.org/
- Pa11y actions tutorial: http://hollsk.co.uk/posts/view/using-actions-in-pa11y
- IRCC rescheduler pa11y file: https://github.com/cds-snc/ircc-rescheduler/blob/master/a11y.test.js
- IRCC rescheduler run commands: https://github.com/cds-snc/ircc-rescheduler/blob/master/package.json#L20-L21