Only show Summary Table edit link if an 'id' is passed in
Previuosly, if the `ifEditable` flag was true, we would put in
"change" links but they all pointed to a hard-coded url.

Since we have to create different urls for each field, I added an
ID field, which is used to build the URL.

And now that we have an id being passed in, we don't actually
need the ifEditable field as a flag anymore.

Easy.