Add the visually hidden legend class to the fieldset not the legend
Scanned the page with axe and it started complaining.
The first child of a fieldset has to be a legend, and putting a
CSS object into the class attribute creates a <style> tag in front
of the element.

So we were getting the following markup:

<fieldset>
  <style>...</style>
  <legend>...</legend>
  ...
</fieldset>

And then failing the validation.

Adding the visuallyhidden markup to the fieldset instead resolves
the issue.