Revamp API
To be able to support future tax slips we're splitting off income into an array of T4s (incl T4a). We're also breaking out personal information in a separate object, and adding the "return" section that has stub information about the return. We'll need to change this in the future to make calculations and generate this object or something of the sort.