Change login to fit new data
While we probably could of just kept the same logic, this removes the  amount of knowledge that the getSessionData function needs to know about the API. We'll pass around the "user" object in full without destructuring it. We can destructure it at a later point when we pass it to the views if need be.

We also change the login text box from name to login and change the name of the data variable to user for clarity.