Added Secret Scanning to CI
I added a step in the CI file that the build fails if secrets are found in the source code. More information about this process can be found here: https://github.com/cds-snc/github-actions/tree/master/seekret. There are false positives which need to be overridden on occasion, like a string that looks a lot like a secret. This is to bring us into compliance with IA-5 (7) of the ITSG 33a.