Initial stab at splitting up address field
Create new properties that match the mock up under an addresses property.
Right now we have home, mailing and mailingSameAsHome. We set mailingSameAsHome  as true if they are the same to remove any doubt in terms of if it's empty or valid or etc.

Also include one private helper function to format the address and 2 getter functions that currently take a user. We may want to revisit that in the future.