Spike out new accordion component
Went to the Aria Practices Authoring Guide and found their accordion
example and then just stole it.

Added a script.js file into the app, but I figured this was coming
down the line anyway.

Should be pretty accessible: handles all the keyboard interactions
that the Aria authoring practices recommends and also adds the corrent
aria attributes.

Created a new page and put the existing accordion on it so that I could
see both accordions side by side. This is very spike, this new page
will be deleted and the new Accordion should replace the existing
code in components/Accordion.js.

But just to keep track of the work I've done, putting a commit in here
so that we can come back to it if need be.

Sources:
- https://www.w3.org/TR/wai-aria-practices-1.1/#accordion
- https://www.w3.org/TR/wai-aria-practices/examples/accordion/accordion.html