Remove previous accordion code from demo page
Since it's kind of tricky to build new components, I created a new
page with both accordions side-by-side so that I could see them
alongside each other.

Realistically, we don't really need the other bits of code, so I've
removed the superfluous parts from the page and now it's just the
markup and CSS for this one component.