Update login and getSessionData method
Updated the login screen to use "#login" instead of "#name" for the
input id.

Also made getSessionData dumber so that the logic is less crazy,
and put the user data into the session.user key.

Original commit message:

While we probably could of just kept the same logic, this removes the  amount of knowledge that the getSessionData function needs to know about the API. We'll pass around the "user" object in full without destructuring it. We can destructure it at a later point when we pass it to the views if need be.

We also change the login text box from name to login and change the name of the data variable to user for clarity.

Original commit:
- c3ed5f934a288a5fbe4b75f62c9a4beed9653cf2
- https://github.com/cds-snc/cra-alpha/commit/c3ed5f934a288a5fbe4b75f62c9a4beed9653cf2